+++ tests/test_typeguard_py38.py
@@ -16,13 +16,15 @@ def test_literal():
 
 @pytest.mark.parametrize('value, total, error_re', [
     ({'x': 6, 'y': 'foo'}, True, None),
-    ({'y': 'foo'}, True, 'the required key "x" is missing for argument "arg"'),
+    ({'y': 'foo'}, True, r'required key\(s\) \("x"\) missing from argument "arg"'),
     ({'x': 6, 'y': 3}, True,
      'type of dict item "y" for argument "arg" must be str; got int instead'),
-    ({'x': 6}, True, 'the required key "y" is missing for argument "arg"'),
+    ({'x': 6}, True, r'required key\(s\) \("y"\) missing from argument "arg"'),
     ({'x': 6}, False, None),
-    ({'x': 'abc'}, False, 'type of dict item "x" for argument "arg" must be int; got str instead')
-], ids=['correct', 'missing_x', 'wrong_y', 'missing_y_error', 'missing_y_ok', 'wrong_x'])
+    ({'x': 'abc'}, False, 'type of dict item "x" for argument "arg" must be int; got str instead'),
+    ({'x': 6, 'foo': 'abc'}, False, r'extra key\(s\) \("foo"\) in argument "arg"'),
+], ids=['correct', 'missing_x', 'wrong_y', 'missing_y_error', 'missing_y_ok', 'wrong_x',
+        'unknown_key'])
 def test_typed_dict(value, total, error_re):
     class DummyDict(TypedDict, total=total):
         x: int